refactor: move execution modes to subdirectory
